fix compile

+++ ziface/iconnmanager.go
@@ -19,8 +19,8 @@ package ziface
 type IConnManager interface {
 	Add(conn IConnection)                   //添加链接
 	Remove(conn IConnection)                //删除连接
-	Get(connID uint32) (IConnection, error) //利用ConnID获取链接
+	Get(connID uint64) (IConnection, error) //利用ConnID获取链接
 	Len() int                               //获取当前连接
 	ClearConn()                             //删除并停止所有链接
-	GetAllConnID() []uint32                 //获取所有连接ID
+	GetAllConnID() []uint64                 //获取所有连接ID
 }
