+++ znet/msghandler.go
@@ -58,7 +58,7 @@ func (mh *MsgHandle) SendMsgToTaskQueue(request ziface.IRequest) {
 	//轮询的平均分配法则
 
 	//得到需要处理此条连接的workerID
-	workerID := request.GetConnection().GetConnID() % mh.WorkerPoolSize
+	workerID := request.GetConnection().GetConnID() % uint64(mh.WorkerPoolSize)
 	//zlog.Ins().DebugF("Add ConnID=%d request msgID=%d to workerID=%d", request.GetConnection().GetConnID(), request.GetMsgID(), workerID)
 	//将请求消息发送给任务队列
 	mh.TaskQueue[workerID] <- request
